requirements.txt
plotly
jupyter
ipykernel
numpy
pandas
google-cloud-secret-manager
gcloud_config_helper
setuptools